Include files that were unstaged
